requirements.txt
amqp==5.0.6
asgiref==3.4.1
ask-sdk-core==1.16.1
ask-sdk-model==1.32.0
ask-sdk-runtime==1.16.1
attrs==21.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.8
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-mjml==0.11.0
django-redis==5.0.0
djangorestframework==3.12.4
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
install==1.3.4
kombu==5.1.0
mccabe==0.6.1
packaging==21.0
pluggy==1.0.0
prompt-toolkit==3.0.21
psycopg==3.0.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.1
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
